Extracted Level model into its own framework
Also moved the Level part of CoreData into this framework. All references to LevelManager and Level have been updated to use the protocols in the Api package of the Level module. Lastly, level factories are now injected into LevelManager, so that it is no longer dependent on the concrete types of Level it supports. This should help with unit testing.